- Adds ProductUsers to existing Campaigns when new events are imported - Adds ProductUsers to existing Campaigns when new ProductUsers are created (if the CampaignAudienceRule is BOTH or NEW) - Improves data parsing for identify call - Accepts reserved identify attributes that are in snake_case
fix date parsing for identify call

product user saves add users to a campaign

improve UX of table loading

adding productUser to campaign based on events

fix typescript issue

+++ packages/ui/services/apollo.ts
@@ -21,8 +21,10 @@ const httpLink = createHttpLink({
 });
 
 const logout = () => {
-  localStorage.clear();
-  window.location.pathname = "/";
+  if (typeof window !== "undefined" && window.localStorage) {
+    localStorage.clear();
+    window.location.pathname = "/";
+  }
 };
 
 const errorLink = onError(({ networkError }) => {
